+++ CHANGES.rst
@@ -7,6 +7,7 @@ in progress
 -----------
 - Fix HTML description metadata
 - Add translation to Espéranto. Thanks, David!
+- Add translation to Russian. Thanks, Elena!
 
 2018-01-17 0.5.0
 ----------------