+++ CHANGES.rst
@@ -8,6 +8,7 @@ in progress
 - Fix HTML description metadata
 - Add translation to Esperanto. Thanks, David!
 - Add translation to Russian. Thanks, Elena!
+- Add translation to Mandarin Chinese. Thanks, Jenny!
 
 2018-01-17 0.5.0
 ----------------