+++ CHANGES.rst
@@ -9,6 +9,7 @@ in progress
 - Add translation to Esperanto. Thanks, David!
 - Add translation to Russian. Thanks, Elena!
 - Add translation to Mandarin Chinese. Thanks, Jenny!
+- Improve typography in "reproduce" section
 
 2018-01-17 0.5.0
 ----------------